+++ test/Microsoft.ApplicationInsights.AspNetCore.Tests/Helpers/CommonMocks.cs
@@ -1,6 +1,7 @@
 ﻿namespace Microsoft.ApplicationInsights.AspNetCore.Tests.Helpers
 {
     using System;
+    using DiagnosticListeners;
     using Microsoft.ApplicationInsights.Channel;
     using Microsoft.ApplicationInsights.Extensibility;
 
@@ -16,6 +17,11 @@
                 InstrumentationKey = InstrumentationKey,
                 TelemetryChannel = new FakeTelemetryChannel { OnSend = onSendCallback }
             });
+        }
+
+        internal static ICorrelationIdLookupHelper MockCorrelationIdLookupHelper()
+        {
+            return new CorrelationIdLookupHelperStub();
         }
     }
 }