+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Changelog
 
+## Version 2.1.0-beta2
+
+- Updated to use the new correlation headers and changed the correlationId to use appId instead of hashed ikey to match other SDKs.
+- Fixed null reference exception for unitialized ILogger.
+- Unit test bug fixes.
+- Upgraded NETStandard.Library dependency to 1.6.1.
+- Updated to reference base SDK 2.4.0-beta2.
+- Included Microsoft.ApplicationInsights.DependencyCollector for .NET Core.
+
 ## Version 2.1.0-beta1
 
 - Bug fixes